% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_cohort_fileset.R
\name{check_cohort}
\alias{check_cohort}
\title{Check cohort file set}
\usage{
check_cohort(cohort_label, cohort_definition)
}
\arguments{
\item{cohort_label}{A string representing the cohort to be checked.}

\item{cohort_definition}{A string with the path to the cohort definition file in xlsx format. The cohort definition is a file with specific information for all our cohorts, that are suitable for the gwascompanion.}
}
\value{
A data.table object containing the results of the check routine.
}
\description{
This function checks that a cohort has all the files required to be used with the gwascompanion.
}
\details{
This function checks that for a given cohort all data files are available in a predefined data format. This is needed for every cohort that's supposed to be input for the gwascompanion. The checks include the availability of a set of files and that some files have specific columns.
}
\examples{
\dontrun{ 
if(interactive()){
  # Check the cohort for its suitability for the gwascompanion
  result <- check_cohort("myCohortData", "path/to/cohort/definition.xlsx")
}
}
}
