% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_qqplot.R
\name{gg_qqplot}
\alias{gg_qqplot}
\title{Create a quantile-quantile plot of p-values with ggplot2.}
\usage{
gg_qqplot(
  pvalues,
  ci = 0.95,
  maf = NULL,
  info = NULL,
  mafcutoff = 0.05,
  infocutoff = 0.8
)
}
\arguments{
\item{pvalues}{Vector of p-values.}

\item{ci}{Size of the confidence interval, 95\% by default.}

\item{maf}{Vector of maf-values, NULL by default.}

\item{info}{Vector of info-values, NULL by default.}

\item{mafcutoff}{cutoff for coloring low maf snps, if maf is not NULL, 0.05 by default.}

\item{infocutoff}{cutoff for different shape of low info snps, if info is not NULL, 0.8 by default.}
}
\value{
A ggplot2 plot.
}
\description{
Assumptions:
\itemize{
\item Expected P values are uniformly distributed.
\item Confidence intervals assume independence between tests.
We expect deviations past the confidence intervals if the tests are
not independent.
For example, in a genome-wide association study, the genotype at any
position is correlated to nearby positions. Tests of nearby genotypes
will result in similar test statistics.
}
}
\examples{
library(ggplot2)
gg_qqplot(runif(1e2)) 
}
\author{
adapted from https://slowkow.com/notes/ggplot2-qqplot/ by Holger
}
