% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSNPTESTsamplefileASdata_table.R
\name{readSNPTESTsamplefileASdata.table}
\alias{readSNPTESTsamplefileASdata.table}
\title{Read SNPTEST Sample File as Data Table}
\usage{
readSNPTESTsamplefileASdata.table(
  snptest_sample_fn,
  makeBinaryOutcomeNumeric12 = T
)
}
\arguments{
\item{snptest_sample_fn}{A string representing the path to the SNPTEST sample file.}

\item{makeBinaryOutcomeNumeric12}{A boolean to specify whether to recode binary outcomes as numeric 1 and 2. Default: T}
}
\value{
A data.table object containing the SNPTEST sample file data with appropriate data types.
}
\description{
This function reads a SNPTEST sample file and converts it into a data.table in R.
}
\details{
This function reads a SNPTEST sample file and converts it into a data.table. It also
converts certain columns to their respective data types and, if specified, recodes binary outcomes
as numeric 1 and 2.
}
\examples{
\dontrun{
if(interactive()){
  # Read SNPTEST sample file and create a data.table object
  snptest_data <- readSNPTESTsamplefileASdata.table("path/to/snptest_sample_file")
}
}
}
