% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeSnptestSamplefile.R
\name{writeSnptestSamplefile}
\alias{writeSnptestSamplefile}
\title{Write SNPTEST Sample File}
\usage{
writeSnptestSamplefile(filename, samplefile, vartypes)
}
\arguments{
\item{filename}{A string representing the path where the SNPTEST sample file will be saved.}

\item{samplefile}{A data.table object containing the SNPTEST sample file data. The first three columns
should be Family-ID, Individual-ID (the latter should only include unique names!), and missing, which can
always be 0, which is typical for imputed data. The column names of the first two columns should be either
"ID1","ID2" or "ID_1","ID_2".}

\item{vartypes}{A character vector specifying the variable types for each column in the sample file. Can be
either "0", "C", "D", "B", or "P". The first three should always be "0","0","0" corresponding to the first
column of the sample file c("ID1","ID2","missing"). 'C' and 'D' are continuous or dichotomous covariates,
'B' and 'P' are binary and continuous outcomes, respectively.}
}
\value{
None. This function writes the SNPTEST sample file to the specified path.
}
\description{
This function writes a SNPTEST sample file with the specified data and variable types.
}
\details{
Write SNPTEST Sample File

This function writes a SNPTEST sample file with the specified data and variable types,
ensuring that the file is formatted correctly for SNPTEST and other genetic programs to read.
}
\examples{
\dontrun{
if(interactive()){
  # Write the SNPTEST sample file
  writeSnptestSamplefile("path/to/output_file", samplefile, vartypes)
}
}
}
